refactor(ui): social connectors util
social connectors util
